@@ -1,19 +1,16 @@
 package com.example.app;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 /**
  * @author https://github.com/ibipn
  *
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest
-public class AppTest {
+class AppTest {
 
   @Test
-  public void contextLoads() {}
+  void contextLoads() {}
 
 }
